Integrate  parts of "CHERI Concepts and Terminology"
This beings addressing the ARC feedback that the ISA spec should describe
the mechanism and defer to other documents for rationale.

As a later step we may want to move the rationale/motivation parts of the
current spec into a new document that can be referenced from the main ISA
specification.