Clarify the added state diagram
Add a "same layout for" prefix to explain why the dashed lines do not
extend through the gap. This is just a limitation of the bytefield
approach that I used with (draw-gap). It would be possible to add more
dashed lines with manual (draw-line) invocations, but the exact
coordinates needed are difficult to compute, so let's just add a
clarification to the text.

Pointed out @heshamelmatary